fix LiveGame rating range selection logic  - replace logic to save gameConfigBuilder instead of gameConfig fix live status notification update and fragment opening
update posted and challenge create successful messages text
improve logic to force stats updates

replace daily games load with sync query for daily current and finished games

fix avg opponent rating float value

fix CCE for DailyHomeTablet portrait mode click

adjust game lessons logic to pass 100% for free moves

update analysis icon for game options

update version 419/3.0.119

+++ src/com/chess/backend/entity/api/stats/GameStatsItem.java
@@ -189,8 +189,8 @@ public class GameStatsItem extends BaseResponseItem<GameStatsItem.Data> {
 			return lowest == null? new BaseRating() : lowest;
 		}
 
-		public float getAverageOpponent() {
-			return average_opponent;
+		public int getAverageOpponent() {
+			return (int) average_opponent;
 		}
 
 		public BestWin getBestWin() {
